marian_charles_mentorship_project/requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
